Fixes #1. Refactor to use PSR-4 autoloading. Setup autolaoding for phpunit tests. Remove loading requirements in favor of use of PSR-4.
