--- EDI/X12/schemas/850_3050.json
@@ -1,1 +1,0 @@
-[{"name": "ST", "min": 1, "max": 1, "children": [{"name": "BEG", "min": 1, "max": 1}, {"name": "NTE", "min": 0, "max": 100}, {"name": "CUR", "min": 0, "max": 1}, {"name": "REF", "min": 0, "max": 99999}, {"name": "PER", "min": 0, "max": 3}, {"name": "TAX", "min": 0, "max": 99999}, {"name": "FOB", "min": 0, "max": 99999}, {"name": "CTP", "min": 0, "max": 99999}, {"name": "PAM", "min": 0, "max": 10}, {"name": "CSH", "min": 0, "max": 5}, {"name": "SAC", "min": 0, "max": 25, "children": [{"name": "CUR", "min": 0, "max": 1}]}, {"name": "ITD", "min": 0, "max": 5}, {"name": "DIS", "min": 0, "max": 20}, {"name": "INC", "min": 0, "max": 1}, {"name": "DTM", "min": 0, "max": 10}, {"name": "LDT", "min": 0, "max": 12}, {"name": "LIN", "min": 0, "max": 5}, {"name": "SI", "min": 0, "max": 2}, {"name": "PID", "min": 0, "max": 200}, {"name": "MEA", "min": 0, "max": 40}, {"name": "PWK", "min": 0, "max": 25}, {"name": "PKG", "min": 0, "max": 200}, {"name": "TD1", "min": 0, "max": 2}, {"name": "TD5", "min": 0, "max": 12}, {"name": "TD3", "min": 0, "max": 12}, {"name": "TD4", "min": 0, "max": 5}, {"name": "MAN", "min": 0, "max": 10}, {"name": "CTB", "min": 0, "max": 5}, {"name": "TXI", "min": 0, "max": 99999}, {"name": "AMT", "min": 0, "max": 99999, "children": [{"name": "AT", "min": 0, "max": 1}, {"name": "REF", "min": 0, "max": 99999}]}, {"name": "N9", "min": 0, "max": 1000, "children": [{"name": "MSG", "min": 0, "max": 1000}]}, {"name": "N1", "min": 0, "max": 200, "children": [{"name": "N2", "min": 0, "max": 2}, {"name": "N3", "min": 0, "max": 2}, {"name": "N4", "min": 0, "max": 99999}, {"name": "NX2", "min": 0, "max": 3}, {"name": "REF", "min": 0, "max": 12}, {"name": "PER", "min": 0, "max": 99999}, {"name": "FOB", "min": 0, "max": 1}, {"name": "TD1", "min": 0, "max": 2}, {"name": "TD5", "min": 0, "max": 12}, {"name": "TD3", "min": 0, "max": 12}, {"name": "TD4", "min": 0, "max": 5}, {"name": "PKG", "min": 0, "max": 200}]}, {"name": "LM", "min": 0, "max": 99999, "children": [{"name": "LQ", "min": 1, "max": 99999}]}, {"name": "SPI", "min": 0, "max": 99999, "children": [{"name": "REF", "min": 0, "max": 5}, {"name": "DTM", "min": 0, "max": 5}, {"name": "MSG", "min": 0, "max": 50}, {"name": "N1", "min": 0, "max": 20, "children": [{"name": "N2", "min": 0, "max": 2}, {"name": "N3", "min": 0, "max": 2}, {"name": "N4", "min": 0, "max": 1}, {"name": "REF", "min": 0, "max": 20}, {"name": "G61", "min": 0, "max": 1}, {"name": "MSG", "min": 0, "max": 50}]}, {"name": "CB1", "min": 0, "max": 99999, "children": [{"name": "REF", "min": 0, "max": 20}, {"name": "DTM", "min": 0, "max": 5}, {"name": "LDT", "min": 0, "max": 1}, {"name": "MSG", "min": 0, "max": 50}]}]}, {"name": "PO1", "min": 1, "max": 100000, "children": [{"name": "SI", "min": 0, "max": 5}, {"name": "CUR", "min": 0, "max": 1}, {"name": "CN1", "min": 0, "max": 1}, {"name": "PO3", "min": 0, "max": 25}, {"name": "CTP", "min": 0, "max": 99999}, {"name": "PAM", "min": 0, "max": 10}, {"name": "MEA", "min": 0, "max": 40}, {"name": "PID", "min": 0, "max": 1000, "children": [{"name": "MEA", "min": 0, "max": 10}]}, {"name": "PWK", "min": 0, "max": 25}, {"name": "PO4", "min": 0, "max": 99999}, {"name": "REF", "min": 0, "max": 99999}, {"name": "PER", "min": 0, "max": 3}, {"name": "SAC", "min": 0, "max": 25, "children": [{"name": "CUR", "min": 0, "max": 1}]}, {"name": "IT8", "min": 0, "max": 1}, {"name": "ITD", "min": 0, "max": 2}, {"name": "DIS", "min": 0, "max": 20}, {"name": "INC", "min": 0, "max": 1}, {"name": "TAX", "min": 0, "max": 99999}, {"name": "FOB", "min": 0, "max": 99999}, {"name": "SDQ", "min": 0, "max": 500}, {"name": "IT3", "min": 0, "max": 5}, {"name": "DTM", "min": 0, "max": 10}, {"name": "TC2", "min": 0, "max": 2}, {"name": "TD1", "min": 0, "max": 1}, {"name": "TD5", "min": 0, "max": 12}, {"name": "TD3", "min": 0, "max": 12}, {"name": "TD4", "min": 0, "max": 5}, {"name": "MAN", "min": 0, "max": 10}, {"name": "MSG", "min": 0, "max": 99999}, {"name": "SPI", "min": 0, "max": 99999}, {"name": "TXI", "min": 0, "max": 99999}, {"name": "SCH", "min": 0, "max": 200, "children": [{"name": "TD1", "min": 0, "max": 2}, {"name": "TD5", "min": 0, "max": 12}, {"name": "TD3", "min": 0, "max": 12}, {"name": "TD4", "min": 0, "max": 5}]}, {"name": "PKG", "min": 0, "max": 200, "children": [{"name": "MEA", "min": 0, "max": 99999}]}, {"name": "LS", "min": 0, "max": 1, "children": [{"name": "LDT", "min": 1, "max": 99999, "children": [{"name": "QTY", "min": 0, "max": 99999}, {"name": "MSG", "min": 0, "max": 1}, {"name": "REF", "min": 0, "max": 3}, {"name": "LM", "min": 0, "max": 99999, "children": [{"name": "LQ", "min": 1, "max": 99999}]}]}, {"name": "LE", "min": 1, "max": 1}]}, {"name": "N9", "min": 0, "max": 1000, "children": [{"name": "MEA", "min": 0, "max": 40}, {"name": "MSG", "min": 0, "max": 1000}]}, {"name": "N1", "min": 0, "max": 200, "children": [{"name": "N2", "min": 0, "max": 2}, {"name": "N3", "min": 0, "max": 2}, {"name": "N4", "min": 0, "max": 1}, {"name": "NX2", "min": 0, "max": 3}, {"name": "REF", "min": 0, "max": 12}, {"name": "PER", "min": 0, "max": 3}, {"name": "FOB", "min": 0, "max": 1}, {"name": "SCH", "min": 0, "max": 200}, {"name": "TD1", "min": 0, "max": 2}, {"name": "TD5", "min": 0, "max": 12}, {"name": "TD3", "min": 0, "max": 12}, {"name": "TD4", "min": 0, "max": 5}, {"name": "PKG", "min": 0, "max": 200}, {"name": "LDT", "min": 0, "max": 99999, "children": [{"name": "MAN", "min": 0, "max": 10}, {"name": "QTY", "min": 0, "max": 5}, {"name": "MSG", "min": 0, "max": 1}, {"name": "REF", "min": 0, "max": 3}]}]}, {"name": "SLN", "min": 0, "max": 1000, "children": [{"name": "SI", "min": 0, "max": 5}, {"name": "PID", "min": 0, "max": 1000}, {"name": "PO3", "min": 0, "max": 104}, {"name": "TC2", "min": 0, "max": 2}, {"name": "SAC", "min": 0, "max": 10, "children": [{"name": "CUR", "min": 0, "max": 1}]}, {"name": "DTM", "min": 0, "max": 10}, {"name": "CTP", "min": 0, "max": 25}, {"name": "PAM", "min": 0, "max": 10}, {"name": "PO4", "min": 0, "max": 1}, {"name": "TAX", "min": 0, "max": 3}, {"name": "N1", "min": 0, "max": 10, "children": [{"name": "N2", "min": 0, "max": 2}, {"name": "N3", "min": 0, "max": 2}, {"name": "N4", "min": 0, "max": 1}, {"name": "NX2", "min": 0, "max": 3}, {"name": "REF", "min": 0, "max": 12}, {"name": "PER", "min": 0, "max": 3}]}]}, {"name": "AMT", "min": 0, "max": 99999, "children": [{"name": "REF", "min": 0, "max": 1}]}, {"name": "LM", "min": 0, "max": 99999, "children": [{"name": "LQ", "min": 1, "max": 99999}]}]}, {"name": "CTT", "min": 1, "max": 1}, {"name": "AMT", "min": 0, "max": 1}, {"name": "SE", "min": 1, "max": 1}]}]
