+++ tests/DocumentTest.php
@@ -1,7 +1,7 @@
 <?php
 
 use PHPUnit\Framework\TestCase;
-use EDI\X12\Document;
+use Pedi\X12\Document;
 
 final class DocumentTest extends TestCase
 {
@@ -9,14 +9,14 @@ final class DocumentTest extends TestCase
     public function testConstructor()
     {
         $test_segment = [1,2,3,4];
-        $doc = new EDI\X12\Document($test_segment);
+        $doc = new Document($test_segment);
         $this->assertInternalType("array", $test_segment);
     }
 
-    public function testDocumentToSring() 
+    public function testDocumentToSring()
     {
         $test_segment = json_decode(file_get_contents(__DIR__ . '/test_files/x12_test_array.json'), true);
-        $doc = new EDI\X12\Document($test_segment);
+        $doc = new Document($test_segment);
         $this->assertInternalType("string", $doc->__toString());
     }
 }
