--- tests/autoload.php
@@ -1,3 +1,0 @@
-<?php
-
-require_once __DIR__.'/../vendor/autoload.php';
